WA-242 Set up redux store in background script
