+++ app/index.js
@@ -6,11 +6,15 @@ import { BrowserRouter } from 'react-router-dom'
 import App from 'components/App'
 import store from 'store'
 
-ReactDOM.render(
-  <Provider store={store}>
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
-  </Provider>,
-  document.getElementById('root')
-)
+store
+  .ready()
+  .then(() => {
+    ReactDOM.render(
+      <Provider store={store}>
+        <BrowserRouter>
+          <App />
+        </BrowserRouter>
+      </Provider>,
+      document.getElementById('root')
+    )
+  })
