+++ package.json
@@ -20,6 +20,7 @@
     "ethereumjs-wallet": "^0.6.0",
     "qr-image": "^3.2.0",
     "react": "^16.2.0",
+    "react-chrome-redux": "^1.5.1",
     "react-dom": "^16.2.0",
     "react-redux": "^5.0.6",
     "react-router": "^4.2.0",
